Add files via upload
Project Description
The Web Developer Portfolio project is a personal web page designed to showcase your skills, projects, and background as a web developer. It serves as a digital resume and portfolio, providing potential employers and clients with a comprehensive overview of your abilities and achievements. The portfolio is built using HTML, CSS, and JavaScript, ensuring a professional and interactive user experience.